requirements.txt
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.2.1
pip==25.1.1
qwen_vl_utils==0.0.11
seaborn==0.13.2
sentence_transformers==4.1.0
torch==2.5.1
tqdm==4.67.1
transformers==4.51.3
